Move persistence from internal to pkg
