+++ kata/internal/statestore/statestore.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/kaleido-io/paladin/kata/internal/cache"
 	"github.com/kaleido-io/paladin/kata/internal/confutil"
-	"github.com/kaleido-io/paladin/kata/internal/persistence"
 	"github.com/kaleido-io/paladin/kata/internal/rpcserver"
+	"github.com/kaleido-io/paladin/kata/pkg/persistence"
 )
 
 type Config struct {
