+++ kata/internal/statestore/statestore_test.go
@@ -21,8 +21,8 @@ import (
 	"testing"
 
 	"github.com/DATA-DOG/go-sqlmock"
-	"github.com/kaleido-io/paladin/kata/internal/persistence"
-	"github.com/kaleido-io/paladin/kata/internal/persistence/mockpersistence"
+	"github.com/kaleido-io/paladin/kata/pkg/persistence"
+	"github.com/kaleido-io/paladin/kata/pkg/persistence/mockpersistence"
 	"github.com/stretchr/testify/assert"
 )
