+++ kata/internal/transaction/submit.go
@@ -21,8 +21,8 @@ import (
 	"github.com/hyperledger/firefly-common/pkg/log"
 	"github.com/kaleido-io/paladin/kata/internal/commsbus"
 	"github.com/kaleido-io/paladin/kata/internal/msgs"
-	"github.com/kaleido-io/paladin/kata/internal/persistence"
 	"github.com/kaleido-io/paladin/kata/internal/transactionstore"
+	"github.com/kaleido-io/paladin/kata/pkg/persistence"
 	transactionsPB "github.com/kaleido-io/paladin/kata/pkg/proto/transaction"
 )
