+++ kata/pkg/persistence/mockpersistence/sqlmock.go
@@ -24,7 +24,7 @@ import (
 	"github.com/DATA-DOG/go-sqlmock"
 	migratedb "github.com/golang-migrate/migrate/v4/database"
 	"github.com/kaleido-io/paladin/kata/internal/confutil"
-	"github.com/kaleido-io/paladin/kata/internal/persistence"
+	"github.com/kaleido-io/paladin/kata/pkg/persistence"
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
 )
