+++ kata/pkg/persistence/mockpersistence/sqlmock_test.go
