+++ kata/pkg/persistence/sqlite.go
