+++ kata/pkg/persistence/sqlite_test.go
